+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### Version 2.0.55
+
+- add SPM migration assistant for iOS projects
+- Support for bun dev server projects
+
 ### Version 2.0.54
 
 - Refactoring project and cleanup of integrations